backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.2.0
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0 